+++ ui-web-react/src/model/actions/race.actions.ts
@@ -1,7 +1,6 @@
 import { AnyAction } from "redux";
-import { RaceItem, RaceItemExt, RacerProfile, RacerResults } from "../types/datatypes";
+import { RaceItem, RaceItemExt } from "../types/datatypes";
 import Optional from "optional-js";
-import { RaceState } from "../types/races.model";
 
 export const RACES_REQUESTED = "RACES_REQUESTED";
 export const RACES_REQUEST_FAILED = "RACES_REQUEST_FAILED";
@@ -10,19 +9,6 @@ export const RACES_LOADED = "RACES_LOADED";
 export const SELECTED_RACE_REQUESTED = "SELECTED_RACE_REQUESTED";
 export const SELECTED_RACE_REQUEST_FAILED = "SELECTED_RACE_REQUEST_FAILED";
 export const SELECTED_RACE_LOADED = "SELECTED_RACE_LOADED";
-
-export const RACE_PARTICIPANTS_UPDATE_REQUESTED = "RACE_PARTICIPANTS_UPDATE_REQUESTED";
-export const RACE_PARTICIPANTS_UPDATED = "RACE_PARTICIPANTS_UPDATED";
-export const RACE_PARTICIPANTS_UPDATE_FAILED = "RACE_PARTICIPANTS_UPDATE_FAILED";
-
-export const RACE_RESULTS_SUBSCRIPTION_STARTED = "RACE_RESULTS_SUBSCRIPTION_STARTED";
-export const RACE_RESULTS_SUBSCRIPTION_STOPPED = "RACE_RESULTS_SUBSCRIPTION_STOPPED";
-export const RACE_RESULTS_SUBSCRIPTION_DATA_RECEIVED = "RACE_RESULTS_SUBSCRIPTION_DATA_RECEIVED";
-export const RACE_RESULTS_SUBSCRIPTION_FAILED = "RACE_RESULTS_SUBSCRIPTION_FAILED";
-
-export const RACE_CHANGE_STATE_REQUESTED = "RACE_CHANGE_STATE_REQUESTED";
-export const RACE_CHANGE_STATE_SUCCESS = "RACE_CHANGE_STATE_SUCCESS";
-export const RACE_CHANGE_STATE_FAILED = "RACE_CHANGE_STATE_FAILED";
 
 export type RacesRequestedAction = AnyAction;
 export type RacesLoadedAction = AnyAction & {
@@ -35,28 +21,6 @@ export type SelectedRaceRequestedAction = AnyAction & {
 export type SelectedRaceLoadedAction = AnyAction & {
   raceItemExt: RaceItemExt;
 };
-export type RaceParticipantsAction = AnyAction & {
-  userUUID: string;
-  raceID: number;
-  itemsAdded: Optional<RacerProfile[]>;
-  itemsRemoved: Optional<RacerProfile[]>;
-};
-export type RaceParticipantsUpdateFailedAction = AnyAction & {
-  raceID: number;
-};
-
-export type RaceResultsSubscriptionAction = AnyAction & {
-  userUUID: string;
-  raceID: number;
-};
-export type RaceResultsSubscriptionDataAction = AnyAction & {
-  data: Optional<RacerResults[]>;
-};
-
-export type RaceChangeStateAction = AnyAction & {
-  raceID: number;
-  state: RaceState;
-};
 
 export const racesRequested = (): RacesRequestedAction => ({
   type: RACES_REQUESTED
@@ -67,28 +31,6 @@ export const racesRequestFailed = (): RacesRequestedAction => ({
 export const racesLoaded = (items: RaceItem[]): RacesLoadedAction => ({
   type: RACES_LOADED,
   items: Optional.of(items)
-});
-const raceParticipantsFactory = (type: string) => (
-  userUUID: string,
-  raceID: number,
-  added: RacerProfile[],
-  removed: RacerProfile[]
-): RaceParticipantsAction => ({
-  type,
-  userUUID,
-  raceID,
-  itemsAdded: Optional.of(added),
-  itemsRemoved: Optional.of(removed)
-});
-export const raceParticipantsUpdateRequested = raceParticipantsFactory(
-  RACE_PARTICIPANTS_UPDATE_REQUESTED
-);
-export const raceParticipantsUpdated = raceParticipantsFactory(RACE_PARTICIPANTS_UPDATED);
-export const raceParticipantsUpdateFailed = (
-  raceID: number
-): RaceParticipantsUpdateFailedAction => ({
-  type: RACE_PARTICIPANTS_UPDATE_FAILED,
-  raceID
 });
 
 export const selectedRaceRequested = (id: number): SelectedRaceRequestedAction => ({
@@ -102,46 +44,4 @@ export const selectedRaceRequestFailed = (id: number): SelectedRaceRequestedActi
 export const selectedRaceLoaded = (raceItemExt: RaceItemExt): SelectedRaceLoadedAction => ({
   type: SELECTED_RACE_LOADED,
   raceItemExt: raceItemExt
-});
-
-const raceResultsSubscriptionFactory = (type: string) => (
-  userUUID: string,
-  raceID: number
-): RaceResultsSubscriptionAction => ({
-  type,
-  userUUID,
-  raceID
-});
-export const raceResultsSubscriptionStarted = raceResultsSubscriptionFactory(
-  RACE_RESULTS_SUBSCRIPTION_STARTED
-);
-export const raceResultsSubscriptionStopped = raceResultsSubscriptionFactory(
-  RACE_RESULTS_SUBSCRIPTION_STOPPED
-);
-export const raceResultsSubscriptionFailed = (): AnyAction => ({
-  type: RACE_RESULTS_SUBSCRIPTION_FAILED
-});
-
-export const raceResultsSubscriptionDataReceived = (
-  data: Optional<RacerResults[]>
-): RaceResultsSubscriptionDataAction => ({
-  type: RACE_RESULTS_SUBSCRIPTION_DATA_RECEIVED,
-  data
-});
-
-export const needToSubscribeToRaceResults = (state: RaceState): boolean =>
-  state === RaceState.STARTED || state === RaceState.STOPPED;
-
-const raceChangeStateFactory = (type: string) => (
-  raceID: number,
-  state: RaceState
-): RaceChangeStateAction => ({
-  type,
-  raceID,
-  state
-});
-export const raceChangeStateRequested = raceChangeStateFactory(RACE_CHANGE_STATE_REQUESTED);
-export const raceChangeStateSuccess = raceChangeStateFactory(RACE_CHANGE_STATE_SUCCESS);
-export const raceChangeStateFailed = (): AnyAction => ({
-  type: RACE_CHANGE_STATE_FAILED
 });
